Only fields with ID are clickable
