+++ src/muuntaja/config/merge/fields8XX.js
@@ -7,9 +7,6 @@
 /* eslint-disable no-unused-vars, array-bracket-newline */
 
 import {Reducers} from '@natlibfi/marc-record-merge';
-import {getDefaultValue} from '../print-to-e/defaults';
-import {MarcRecord} from '@natlibfi/marc-record';
-import {validationOff} from '../common';
 
 //-------------------------------------------------------------------------
 // Sarjalisäkirjauskentät 80x - 830:
@@ -28,30 +25,12 @@ const fieldsE2P = [
   //copy('830'),
 ].join('|');
 
-
 export function merge8XX(opts) {
-
-  function fillDefault(tag) {
-    return (baseRecord, source) => {
-      const base = new MarcRecord(baseRecord, validationOff);
-      const field = getDefaultValue(tag, opts);
-      base.insertField(field);
-      return base;
-    };
-  }
-
-  function fieldsFenni() {
-    if (opts.profile === 'FENNI') {
-      return [fillDefault('856/FENNI')];
-    }
-    return [];
-  }
 
   if (opts.type === 'e2p') {
     return [
-      Reducers.copy({tagPattern: fieldsE2P}),
-      ...fieldsFenni()
+      Reducers.copy({tagPattern: fieldsE2P})
     ];
   }
-  return [...fieldsFenni()];
+  return [];
 }
