+++ src/muuntaja/config/merge/fields9XX.js
@@ -7,9 +7,6 @@
 /* eslint-disable no-unused-vars, array-bracket-newline */
 
 import {Reducers} from '@natlibfi/marc-record-merge';
-import {getDefaultValue} from '../print-to-e/defaults';
-import {MarcRecord} from '@natlibfi/marc-record';
-import {validationOff} from '../common';
 
 //9XX ei oteta
 
@@ -25,27 +22,10 @@ const fieldsE2P = [
 
 export function merge9XX(opts) {
 
-  function fillDefault(tag) {
-    return (baseRecord, source) => {
-      const base = new MarcRecord(baseRecord, validationOff);
-      const field = getDefaultValue(tag, opts);
-      base.insertField(field);
-      return base;
-    };
-  }
-
-  function fieldsFenni() {
-    if (opts.profile === 'FENNI') {
-      return [fillDefault('901/FENNI')];
-    }
-    return [];
-  }
-
   if (opts.type === 'e2p') {
     return [
-      Reducers.copy({tagPattern: fieldsE2P}),
-      ...fieldsFenni()
+      Reducers.copy({tagPattern: fieldsE2P})
     ];
   }
-  return [...fieldsFenni()];
+  return [];
 }
